pyproject.toml
[project]
name = "deezer-dl-bot"
description = "Deezer-DL Bot for Jellyfin"
readme = "README.md"
version = "0.0.6"
dependencies = [
  "deezer_dl @ git+ssh://git@github.com/MaximeSahuc/deezer-dl-cli",
  "pyyaml",
]
authors = [
  {name = "Maxime Sahuc"},
]

[project.urls]
Repository = "https://github.com/MaximeSahuc/deezer-dl-cli.git"

[project.scripts]
deezer-dl-bot = "main:main"